Rewrite profile picture view and remove a few unused classes
