+++ app/src/main/java/org/thoughtcrime/securesms/contacts/UserView.kt
@@ -53,7 +53,7 @@ class UserView : LinearLayout {
             return contact?.displayName(Contact.ContactContext.REGULAR) ?: publicKey
         }
         val address = user.address.serialize()
-        binding.profilePictureView.update(user)
+        binding.profilePictureView.load(user)
         binding.actionIndicatorImageView.setImageResource(R.drawable.ic_baseline_edit_24)
         binding.nameTextView.text = if (user.isGroupRecipient) user.name else getUserDisplayName(address)
         when (actionIndicator) {
@@ -85,7 +85,6 @@ class UserView : LinearLayout {
     }
 
     fun unbind() {
-        binding.profilePictureView.recycle()
     }
     // endregion
 }
