+++ app/src/main/java/org/thoughtcrime/securesms/home/ConversationView.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.content.res.Resources
 import android.graphics.Typeface
 import android.graphics.drawable.ColorDrawable
-import android.text.TextUtils
 import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.View
@@ -128,11 +127,10 @@ class ConversationView : LinearLayout {
             thread.isRead -> binding.statusIndicatorImageView.setImageResource(R.drawable.ic_filled_circle_check)
             else -> binding.statusIndicatorImageView.setImageResource(R.drawable.ic_circle_check)
         }
-        binding.profilePictureView.update(thread.recipient)
+        binding.profilePictureView.load(thread.recipient)
     }
 
     fun recycle() {
-        binding.profilePictureView.recycle()
     }
 
     private fun getTitle(recipient: Recipient): String? = when {
