+++ app/src/main/java/org/thoughtcrime/securesms/home/UserDetailsBottomSheet.kt
@@ -55,8 +55,7 @@ class UserDetailsBottomSheet: BottomSheetDialogFragment() {
         val recipient = Recipient.from(requireContext(), Address.fromSerialized(publicKey), false)
         val threadRecipient = threadDb.getRecipientForThreadId(threadID) ?: return dismiss()
         with(binding) {
-            profilePictureView.publicKey = publicKey
-            profilePictureView.update(recipient)
+            profilePictureView.load(recipient)
             nameTextViewContainer.visibility = View.VISIBLE
             nameTextViewContainer.setOnClickListener {
                 if (recipient.isOpenGroupInboxRecipient || recipient.isOpenGroupOutboxRecipient) return@setOnClickListener
