+++ app/src/main/java/org/thoughtcrime/securesms/messagerequests/MessageRequestView.kt
@@ -49,12 +49,11 @@ class MessageRequestView : LinearLayout {
         binding.snippetTextView.text = snippet
 
         post {
-            binding.profilePictureView.update(thread.recipient)
+            binding.profilePictureView.load(thread.recipient)
         }
     }
 
     fun recycle() {
-        binding.profilePictureView.recycle()
     }
 
     private fun getUserDisplayName(recipient: Recipient): String? {
