+++ app/src/main/java/org/thoughtcrime/securesms/reactions/ReactionRecipientsAdapter.java
@@ -13,7 +13,6 @@ import org.session.libsession.messaging.utilities.AccountId;
 import org.thoughtcrime.securesms.components.ProfilePictureView;
 import org.thoughtcrime.securesms.components.emoji.EmojiImageView;
 import org.thoughtcrime.securesms.database.model.MessageId;
-import com.bumptech.glide.Glide;
 
 import java.util.Collections;
 import java.util.List;
@@ -154,7 +153,7 @@ final class ReactionRecipientsAdapter extends RecyclerView.Adapter<ReactionRecip
         callback.onRemoveReaction(reaction.getBaseEmoji(), messageId, reaction.getTimestamp());
       });
 
-      this.avatar.update(reaction.getSender());
+      this.avatar.load(reaction.getSender(), true);
 
       if (reaction.getSender().isLocalNumber()) {
         this.recipient.setText(R.string.ReactionsRecipientAdapter_you);
@@ -170,7 +169,6 @@ final class ReactionRecipientsAdapter extends RecyclerView.Adapter<ReactionRecip
     }
 
     void unbind() {
-      avatar.recycle();
     }
 
   }
